+++ app/Http/Controllers/Backsite/RoleController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 // use library here
 use Illuminate\Support\Facades\Storage;
 use Symfony\Component\HttpFoundation\Response;
+use Illuminate\Support\Facades\Gate;
 
 // user request
 use App\Http\Requests\Role\StoreRoleRequest;
@@ -44,6 +45,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('role_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $role = Role::orderBy('created_at', 'desc')->get();
 
@@ -78,6 +80,8 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        abort_if(Gate::denies('role_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         // need more notes here
         $role->load('permission');
 
@@ -89,6 +93,8 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        abort_if(Gate::denies('role_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         // need more notes here
         $permission = Permission::all();
         $role->load('permission');
@@ -101,7 +107,7 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, Role $role)
     {
-
+        // need more notes here
         $role->update($request->all());
         $role->permission()->sync($request->input('permission', []));
 
@@ -114,6 +120,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        abort_if(Gate::denies('role_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $role->forceDelete();
 
         alert()->success('Success Message', 'Successfully deleted role');
